Reverting eigen matrices back to double

+++ src/LinearRegression.cpp
@@ -16,7 +16,7 @@ LinearRegression::LinearRegression(Dataset &dataset, std::vector<Class> &classes
 }
 
 void LinearRegression::train() {
-    MatrixXf X(dataset.size(), dataset.dimension() + 1);
+    MatrixXd X(dataset.size(), dataset.dimension() + 1);
 
     // Reading data and transferring to observation matrix
     for (int i = 0; i < dataset.size(); ++i) {
@@ -25,7 +25,7 @@ void LinearRegression::train() {
         X(i, 0) = 1.0;
     }
 
-    MatrixXf Z(dataset.size(), (int) N_CLASSES);
+    MatrixXd Z(dataset.size(), (int) N_CLASSES);
     Z.setZero();
 
     for (int i = 0; i < dataset.size(); ++i) {
@@ -36,18 +36,18 @@ void LinearRegression::train() {
         }
     }
 
-    MatrixXf Xt = X.transpose();
+    MatrixXd Xt = X.transpose();
 
     B = (Xt * X).inverse() * Xt * Z;
 }
 
 Class LinearRegression::classify(Token &token) {
-    RowVectorXf vec(dataset.dimension() + 1);
+    RowVectorXd vec(dataset.dimension() + 1);
     for (int i = 1; i < dataset.dimension() + 1; ++i)
         vec[i] = token[i - 1];
     vec[0] = 1.0;
 
-    RowVectorXf pred = vec * B;
+    RowVectorXd pred = vec * B;
 
     auto clf = (Class) 0;
     for (int i = 0; i < N_CLASSES; ++i)
@@ -58,12 +58,12 @@ Class LinearRegression::classify(Token &token) {
 }
 
 float LinearRegression::classificationProbability(Class c, Token &token) {
-    RowVectorXf vec(dataset.dimension() + 1);
+    RowVectorXd vec(dataset.dimension() + 1);
     for (int i = 1; i < dataset.dimension() + 1; ++i)
         vec[i] = token[i - 1];
     vec[0] = 1.0;
 
-    RowVectorXf pred = vec * B;
+    RowVectorXd pred = vec * B;
 
     return pred[(int) c];
 }
